requirements.txt
torch
torchvision
datasets
Pillow
numpy
matplotlib
scikit-learn
